+++ src/main.js
@@ -6,7 +6,7 @@ import {generateMenu} from "./view/menu.js";
 import {generateMoreButton} from "./view/more-button.js";
 import {generateSort} from "./view/sort.js";
 import {generateTasksContainer} from "./view/task-container.js";
-import {CARDS_COUNT, RenderPosition} from "./const.js";
+import {CARDS_COUNT, RenderPosition} from "./utils/const.js";
 import {render} from "./utils/index.js";
 
 const mainContainer = document.querySelector(`.main`);
